Do not remember convergance_strategy for every machine
Before we were remembering the state of the convergance_strategy for
all the machine resources. That causes that if the user changed some
attributes like the `chef_server_url` to point to, it was not taking
effect.

This commit fixes that memory that we do not need to have. :smile:

Bump Version to 0.8.4